Center text, import Tesseract, upload images to server
Center the text on the webpage.  Import the tesseract NuGet package and determine how to use it.  Running into difficulty of Tesseract getting a good file path to process images.  Upload images to server for processing with tesseract.  Having issues waiting for file path data to get saved to object before tesseract component renders on page.